# backend/docs/swagger.yaml
definitions:
...
  ErrorResponseModel:
    properties:
      error:
        type: string
    type: object
...
  WindowModelCharacteristics:
    properties:
      chambers:
        type: string
      executions:
        items:
          type: string
        type: array
      falz_height:
        type: integer
      frame_sash_height:
        type: integer
      glass_type:
        type: string
      profile:
        type: string
      seal_colors:
        items:
          type: string
        type: array
      seal_material:
        items:
          type: string
        type: array
      thermal_resistance:
        type: number
      width:
        type: integer
    type: object
  WindowModelReadObject:
    properties:
      characteristics:
        $ref: '#/definitions/WindowModelCharacteristics'
      id:
        type: string
      large_image_path:
        type: string
      material_id:
        description: Эти поля останутся для других операций
        type: string
      material_name:
        description: Новые поля для хранения имен
        type: string
      medium_image_path:
        type: string
      name:
        type: string
      small_image_path:
        type: string
      system_id:
        description: Эти поля останутся для других операций
        type: string
      system_name:
        description: Новые поля для хранения имен
        type: string
      type_id:
        description: Эти поля останутся для других операций
        type: string
      type_name:
        description: Новые поля для хранения имен
        type: string
    type: object
  WindowTypesCreate:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  WindowTypesRead:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /window:
    get:
      consumes:
      - application/json
      description: Retrieve all window types from the database
      produces:
      - application/json
      responses:
        "200":
          description: List of window types
          schema:
            items:
              $ref: '#/definitions/WindowTypesRead'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Get all window types
      tags:
      - window
    post:
      consumes:
      - application/json
      description: Create a new window type with the provided data
      parameters:
      - description: Window type data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/WindowTypesCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Object created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Create a new window type
      tags:
      - window
  /window/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a window type from the database using its unique identifier
      parameters:
      - description: ID of the window type to delete
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Window type deleted successfully
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Delete a window type
      tags:
      - window
    get:
      consumes:
      - application/json
      description: Retrieve window type details from the database using its unique
        identifier
      parameters:
      - description: ID of the window type to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Window type object
          schema:
            $ref: '#/definitions/WindowTypesRead'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Get window type by ID
      tags:
      - window
    put:
      consumes:
      - application/json
      description: Update window type data by ID
      parameters:
      - description: ID of the window type to update
        in: path
        name: id
        required: true
        type: string
      - description: Updated window type data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/WindowTypesCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Object updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Update window type
      tags:
      - window
  /window/model:
    get:
      consumes:
      - application/json
      description: Retrieve all window models from the database
      produces:
      - application/json
      responses:
        "200":
          description: List of window models
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Get all window models
      tags:
      - window-model
    post:
      consumes:
      - multipart/form-data
      description: Create a new window model with images and characteristics
      parameters:
      - description: Large image file (jpeg/png)
        in: formData
        name: large_image
        required: true
        type: file
      - description: Medium image file (jpeg/png)
        in: formData
        name: medium_image
        required: true
        type: file
      - description: Small image file (jpeg/png)
        in: formData
        name: small_image
        required: true
        type: file
      - description: Window model data in JSON format
        in: formData
        name: object_data
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Window model created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Create a new window model
      tags:
      - window-model
  /window/model/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a window model from the database using its unique identifier
      parameters:
      - description: ID of the window model to delete
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: window model deleted!
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Delete a window model
      tags:
      - window-model
    get:
      consumes:
      - application/json
      description: Retrieve window model details from the database using its unique
        identifier
      parameters:
      - description: ID of the window model to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Window model object
          schema:
            $ref: '#/definitions/WindowModelReadObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Get window model by ID
      tags:
      - window-model
    put:
      consumes:
      - multipart/form-data
      description: Update window model characteristics by ID
      parameters:
      - description: ID of the window model to update
        in: path
        name: id
        required: true
        type: string
      - description: Updated window model characteristics in JSON format
        in: formData
        name: object_data
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Window model updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Update window model characteristics
      tags:
      - window-model
swagger: "2.0"